pyproject.toml
[tool.poetry]
name = "ds_migrator"
version = "0.0.1"
description = ""
authors = ["Alex Jackson <alex_jackson@trendmicro.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.7"
urllib3 = "^1.26.4"
six = "^1.15.0"
certifi = "^2020.12.5"
python-dateutil = "^2.8.1"
deep-security-api = "^20.0.60"
prompt-toolkit = "^3.0.18"

[tool.poetry.dev-dependencies]
pylint = "^2.7.2"
black = "^20.8b1"

[tool.poetry.scripts]
dsmg = 'ds_migrator.__main__:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
